requirements.txt
flask
pypdf
faiss-cpu
google-genai
python-dotenv
numpy
gunicorn
